requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
openpyxl>=3.1.5
tenacity>=9.0.0
rich>=13.7.1
pydantic>=2.8.2
